+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Upcoming
 
 * Improve boost overview (now shows how much sp need to get invested to make a place boostable and safe)
+* Renamed 'rewards' to 'incidents'
 
 # 0.6.0
 * Include advanced tactics in calculated attack and defense bonus